Refactor: divide post component for maintainability .
